fix: omit employee files and add code to user the  files
