+++ backend/http/auth/register.http
@@ -2,8 +2,8 @@ POST http://localhost:3001/auth/register HTTP/1.1
 Content-Type: application/json
 
 {
-	"email": "goli@d.com",
+	"email": "sara@d.com",
     "password": "0000",
-    "name": "goli Doe",
-    "organizationSlug": "hi goli's Company"
+    "name": "sara Doe",
+    "organizationSlug": "hi sara's Company"
 }
