+++ backend/http/employee.http/deleteEmployee.http
@@ -1,3 +1,3 @@
-DELETE http://localhost:3001/employee/9 HTTP/1.1
+DELETE http://localhost:3001/users/4 HTTP/1.1
 Content-Type: application/json
-Authorization: bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImdvbGlAZC5jb20iLCJzdWIiOjEsIm5hbWUiOiJnb2xpIERvZSIsInJvbGUiOnsiaWQiOjEsIm5hbWUiOiJPd25lciIsImNyZWF0ZWRBdCI6IjIwMjItMTAtMjNUMDM6MTI6NTIuOTAxWiIsInVwZGF0ZWRBdCI6IjIwMjItMTAtMjNUMDM6MTI6NTIuOTAxWiJ9LCJpYXQiOjE2NjY1MDc0NTYsImV4cCI6MTY2OTA5OTQ1Nn0.nKBDHT7LqOwqZtvB1Y2FYZlX9wLk1tiArujMaR4PEzs
+Authorization: bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImdvbGlAZC5jb20iLCJzdWIiOjEsIm5hbWUiOiJnb2xpIERvZSIsInJvbGUiOnsiaWQiOjEsIm5hbWUiOiJPd25lciIsImNyZWF0ZWRBdCI6IjIwMjItMTAtMjRUMDY6MjU6NTEuMTEzWiIsInVwZGF0ZWRBdCI6IjIwMjItMTAtMjRUMDY6MjU6NTEuMTEzWiJ9LCJpYXQiOjE2NjY2MDk0MzAsImV4cCI6MTY2OTIwMTQzMH0.OPp31gBW2WiH984-UbHetAhQLMGnUrql2_ZzLMdK3rw
