+++ backend/http/employee.http/getOneEmployee.http
@@ -1,2 +1,2 @@
-GET http://localhost:3001/employee/2 HTTP/1.1
-Authorization: bearer  eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImdvbGlAZC5jb20iLCJzdWIiOjEsIm5hbWUiOiJnb2xpIERvZSIsInJvbGUiOnsiaWQiOjEsIm5hbWUiOiJPd25lciIsImNyZWF0ZWRBdCI6IjIwMjItMTAtMjRUMDE6MDQ6MzYuMjQzWiIsInVwZGF0ZWRBdCI6IjIwMjItMTAtMjRUMDE6MDQ6MzYuMjQzWiJ9LCJpYXQiOjE2NjY1ODYxNDgsImV4cCI6MTY2OTE3ODE0OH0.B030RjFFmy4DitPpO9BaVzWPSRTHYqnZVoKzNXcil1k
+GET http://localhost:3001/users/3 HTTP/1.1
+Authorization: bearer  eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InNhcmFAZC5jb20iLCJzdWIiOjUsIm5hbWUiOiJzYXJhIERvZSIsInJvbGUiOnsiaWQiOjEsIm5hbWUiOiJPd25lciIsImNyZWF0ZWRBdCI6IjIwMjItMTAtMjRUMDY6MjU6NTEuMTEzWiIsInVwZGF0ZWRBdCI6IjIwMjItMTAtMjRUMDY6MjU6NTEuMTEzWiJ9LCJpYXQiOjE2NjY2MDg0MTEsImV4cCI6MTY2OTIwMDQxMX0.4hE02QIw998LqTuN0OdED6X3F48KV_AVCFqZGpmMcGg
