+++ backend/http/employee.http/readAllEmployeeByOrganization.http
@@ -1,3 +1,3 @@
-GET http://localhost:3001/employee HTTP/1.1
+GET http://localhost:3001/users HTTP/1.1
 Content-Type: application/json
-Authorization: bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImdvbGlAZC5jb20iLCJzdWIiOjEsIm5hbWUiOiJnb2xpIERvZSIsInJvbGUiOnsiaWQiOjEsIm5hbWUiOiJPd25lciIsImNyZWF0ZWRBdCI6IjIwMjItMTAtMjRUMDE6MDQ6MzYuMjQzWiIsInVwZGF0ZWRBdCI6IjIwMjItMTAtMjRUMDE6MDQ6MzYuMjQzWiJ9LCJpYXQiOjE2NjY1ODYxNDgsImV4cCI6MTY2OTE3ODE0OH0.B030RjFFmy4DitPpO9BaVzWPSRTHYqnZVoKzNXcil1k
+Authorization: bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImdvbGlAZC5jb20iLCJzdWIiOjEsIm5hbWUiOiJnb2xpIERvZSIsInJvbGUiOnsiaWQiOjEsIm5hbWUiOiJPd25lciIsImNyZWF0ZWRBdCI6IjIwMjItMTAtMjRUMDY6MjU6NTEuMTEzWiIsInVwZGF0ZWRBdCI6IjIwMjItMTAtMjRUMDY6MjU6NTEuMTEzWiJ9LCJpYXQiOjE2NjY2MDk0MzAsImV4cCI6MTY2OTIwMTQzMH0.OPp31gBW2WiH984-UbHetAhQLMGnUrql2_ZzLMdK3rw
