+++ backend/http/employee.http/updateEmployee.http
@@ -1,4 +1,4 @@
-PUT http://localhost:3001/employee/2 HTTP/1.1
+PUT http://localhost:3001/users/2 HTTP/1.1
 Content-Type: application/json
 Authorization: bearer  eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImdvbGlAZC5jb20iLCJzdWIiOjEsIm5hbWUiOiJnb2xpIERvZSIsInJvbGUiOnsiaWQiOjEsIm5hbWUiOiJPd25lciIsImNyZWF0ZWRBdCI6IjIwMjItMTAtMjNUMDM6MTI6NTIuOTAxWiIsInVwZGF0ZWRBdCI6IjIwMjItMTAtMjNUMDM6MTI6NTIuOTAxWiJ9LCJpYXQiOjE2NjY1MDc0NTYsImV4cCI6MTY2OTA5OTQ1Nn0.nKBDHT7LqOwqZtvB1Y2FYZlX9wLk1tiArujMaR4PEzs
