+++ backend/http/invite/createNewInvite.http
@@ -1,12 +1,12 @@
 POST http://localhost:3001/invites HTTP/1.1
 Content-Type: application/json
-Authorization: bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImdvbGlAZC5jb20iLCJzdWIiOjQsIm5hbWUiOiJnb2wgRG9lIiwiaWF0IjoxNjY2MDg3MjAzLCJleHAiOjE2Njg2NzkyMDN9.zj_DHLIb0ULgSlhNdMWfDo_ZZiKbVgWkBsran1kpn1w
+Authorization: bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InNhcmFAZC5jb20iLCJzdWIiOjUsIm5hbWUiOiJzYXJhIERvZSIsInJvbGUiOnsiaWQiOjEsIm5hbWUiOiJPd25lciIsImNyZWF0ZWRBdCI6IjIwMjItMTAtMjRUMDY6MjU6NTEuMTEzWiIsInVwZGF0ZWRBdCI6IjIwMjItMTAtMjRUMDY6MjU6NTEuMTEzWiJ9LCJpYXQiOjE2NjY2MDg0MTEsImV4cCI6MTY2OTIwMDQxMX0.4hE02QIw998LqTuN0OdED6X3F48KV_AVCFqZGpmMcGg
 
 {
 	"invites": [
 		{ 
-			"name": "banana",
-			"email": "banana2@yahoo.com.com"
+			"name": "saba rad",
+			"email": "sabarad@gmail.com"
 		}
 	]
 }
