+++ backend/http/invite/registerUserByToken.http
@@ -1,8 +1,7 @@
-POST http://localhost:3001/invites/16a24e6aad4982f2feeebb404ebd08c4fe7b5df0099f0acade2c1b06503c17a514a696be1ed746f190504f89764ea5ec HTTP/1.1
+POST http://localhost:3001/invites/0b64d82c66448b9df76a0dc1c9c646d45b66a0a34b21784f8b39d8689be926f7c1868e1e1ab41c995acf30370e60cc3b HTTP/1.1
 Content-Type: application/json
-Authorization: bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImdvbGlAZC5jb20iLCJzdWIiOjEsIm5hbWUiOiJnb2xpIERvZSIsInJvbGUiOnsiaWQiOjEsIm5hbWUiOiJPd25lciIsImNyZWF0ZWRBdCI6IjIwMjItMTAtMjNUMDM6MTI6NTIuOTAxWiIsInVwZGF0ZWRBdCI6IjIwMjItMTAtMjNUMDM6MTI6NTIuOTAxWiJ9LCJpYXQiOjE2NjY1MDc0NTYsImV4cCI6MTY2OTA5OTQ1Nn0.nKBDHT7LqOwqZtvB1Y2FYZlX9wLk1tiArujMaR4PEzs
 
 {
-	"name": "banana",
+	"name": "saba rad",
     "password": "0000"
 }
