+++ backend/src/app.module.ts
@@ -36,8 +36,6 @@ import { Log } from './entities/log.entity';
 import { StatusService } from './services/contact/status.service';
 import { NoteController } from './controllers/note.controller';
 import { NoteService } from './services/note.service';
-import { EmployeeService } from './services/employee.service';
-import { EmployeeController } from './controllers/employee.controller';
 import { SearchService } from './services/search.service';
 import { SearchController } from './controllers/search.controller';
 import { ContactStatus } from './entities/contact/contactStatus.entity';
@@ -124,7 +122,6 @@ const entitiesToAdd = [
     OrganizationController,
     InviteController,
     NoteController,
-    EmployeeController,
     SearchController,
   ],
   providers: [
@@ -142,7 +139,6 @@ const entitiesToAdd = [
     RequestContextService,
     StatusService,
     NoteService,
-    EmployeeService,
     SearchService,
     LogService,
   ],
