+++ backend/src/interfaces/user.service.interface.ts
@@ -1,5 +1,7 @@
+import { Organization } from 'src/entities/organization.entity';
 import { User } from 'src/entities/user.entity';
 import { NewUser } from 'src/utils/types';
+import { DeepPartial } from 'typeorm';
 
 export interface IAddUserPayload {
   user: NewUser;
@@ -29,4 +31,21 @@ export interface IExistAndFindByEmailPayload {
 
 export interface IFindUserWithOrganizationPayload {
   email: string;
+}
+
+export interface IReadAllByOrganization {
+  organization: Organization;
+}
+
+export interface IReadOneById {
+  id: number;
+}
+
+export interface IUpdateOneByIdPayload {
+  id: number;
+  employee: DeepPartial<User>;
+}
+
+export interface IDeleteOneByIdPayload {
+  id: number;
 }
